requirements.txt
Django~=5.1.3
djangorestframework~=3.15.2
channels~=4.2.0
pip~=24.3.1
attrs~=24.2.0
wheel~=0.45.1
cryptography~=43.0.3
pillow~=10.2.0
docutils~=0.21.2
jinja2~=3.1.4
asgiref~=3.8.1
sqlparse~=0.5.2
mysqlclient~=2.1.1
psycopg2~=2.9.9
colorama~=0.4.6
PyYAML~=6.0
numpy~=1.26.4
requests~=2.28.1
setuptools~=70.0.0
packaging~=23.2
pyparsing~=3.1.4
future~=1.0.0
uritemplate~=4.1.1
inflection~=0.5.1
Markdown~=3.7
pygments~=2.18.0
pytz~=2022.7
googletrans==4.0.0-rc1